requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
discord-components==2.1.2
discord-py-interactions==4.1.0
discord-py-slash-command==3.0.3
discord-ui==5.1.6
discord.py==1.7.3
Flask==2.1.2
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gspread==5.3.2
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.0
orjson==3.6.7
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygsheets==2.0.5
pyparsing==3.0.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2